fix(api): gate versus_router behind admin
Versus endpoints (/api/versus/*) had no auth dependencies, leaving
eval data readable by any signed-in user while traces and stats had
already been gated behind admin. Add a router-level
dependencies=[Depends(require_admin)] to bring versus in line.

To break the import cycle (app.py imports versus_router, versus_router
needs require_admin from app.py), move _get_admin_db and require_admin
from app.py into auth.py. Function bodies are unchanged; app.py imports
them back so existing call sites are untouched.
